requirements.txt
axiom-axle>=0.8.0
openai>=1.60.0
PyYAML>=6.0.0
